refactor: 영화 예매 화면 네이밍 수정
- Reservation -> ReservationDetail
