+++ app/src/main/res/layout/activity_reservation.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main"
-    tools:context=".presentation.view.reservation.reservation.ReservationActivity">
+    tools:context=".presentation.view.reservation.detail.ReservationDetailActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
